requirements(2).txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-multipart==0.0.6
httpx==0.26.0
langgraph==0.0.55
langchain-core==0.1.52
langchain-community==0.0.24
dashscope==1.14.1
neo4j==5.18.0
pymilvus==2.4.0
streamlit==1.30.0
requests==2.31.0
python-dotenv==1.0.0
